+++ rocketpy/sensors/gyroscope.py
@@ -296,3 +296,28 @@ class Gyroscope(InertialSensor):
             file_format=file_format,
             data_labels=("t", "wx", "wy", "wz"),
         )
+
+    def to_dict(self, **kwargs):
+        data = super().to_dict(**kwargs)
+        data.update({"acceleration_sensitivity": self.acceleration_sensitivity})
+        return data
+
+    @classmethod
+    def from_dict(cls, data):
+        return cls(
+            sampling_rate=data["sampling_rate"],
+            orientation=data["orientation"],
+            measurement_range=data["measurement_range"],
+            resolution=data["resolution"],
+            noise_density=data["noise_density"],
+            noise_variance=data["noise_variance"],
+            random_walk_density=data["random_walk_density"],
+            random_walk_variance=data["random_walk_variance"],
+            constant_bias=data["constant_bias"],
+            operating_temperature=data["operating_temperature"],
+            temperature_bias=data["temperature_bias"],
+            temperature_scale_factor=data["temperature_scale_factor"],
+            cross_axis_sensitivity=data["cross_axis_sensitivity"],
+            acceleration_sensitivity=data["acceleration_sensitivity"],
+            name=data["name"],
+        )