+++ rocketpy/simulation/flight.py
@@ -1,13 +1,13 @@
 # pylint: disable=too-many-lines
-import json
 import math
 import warnings
 from copy import deepcopy
 from functools import cached_property
 
 import numpy as np
-import simplekml
 from scipy.integrate import BDF, DOP853, LSODA, RK23, RK45, OdeSolver, Radau
+
+from rocketpy.simulation.flight_data_exporter import FlightDataExporter
 
 from ..mathutils.function import Function, funcify_method
 from ..mathutils.vector_matrix import Matrix, Vector
@@ -15,6 +15,7 @@ from ..plots.flight_plots import _FlightPlots
 from ..prints.flight_prints import _FlightPrints
 from ..tools import (
     calculate_cubic_hermite_coefficients,
+    deprecated,
     euler313_to_quaternions,
     find_closest,
     find_root_linear_interpolation,
@@ -143,7 +144,7 @@ class Flight:
     Flight.heading : int, float
         Launch heading angle relative to north given in degrees.
     Flight.initial_solution : list
-        List defines initial condition - [tInit, x_init,
+        List defines initial condition - [t_initial, x_init,
         y_init, z_init, vx_init, vy_init, vz_init, e0_init, e1_init,
         e2_init, e3_init, w1_init, w2_init, w3_init]
     Flight.t_initial : int, float
@@ -155,12 +156,6 @@ class Flight:
         Current integration time.
     Flight.y : list
         Current integration state vector u.
-    Flight.post_processed : bool
-        Defines if solution data has been post processed.
-    Flight.initial_solution : list
-        List defines initial condition - [tInit, x_init,
-        y_init, z_init, vx_init, vy_init, vz_init, e0_init, e1_init,
-        e2_init, e3_init, w1_init, w2_init, w3_init]
     Flight.out_of_rail_time : int, float
         Time, in seconds, in which the rocket completely leaves the
         rail.
@@ -546,7 +541,7 @@ class Flight:
         rtol : float, array, optional
             Maximum relative error tolerance to be tolerated in the
             integration scheme. Can be given as array for each
-            state space variable. Default is 1e-3.
+            state space variable. Default is 1e-6.
         atol : float, optional
             Maximum absolute error tolerance to be tolerated in the
             integration scheme. Can be given as array for each
@@ -767,7 +762,22 @@ class Flight:
                         callbacks = [
                             lambda self, parachute_cd_s=parachute.cd_s: setattr(
                                 self, "parachute_cd_s", parachute_cd_s
-                            )
+                            ),
+                            lambda self, parachute_radius=parachute.radius: setattr(
+                                self, "parachute_radius", parachute_radius
+                            ),
+                            lambda self, parachute_height=parachute.height: setattr(
+                                self, "parachute_height", parachute_height
+                            ),
+                            lambda self, parachute_porosity=parachute.porosity: setattr(
+                                self, "parachute_porosity", parachute_porosity
+                            ),
+                            lambda self,
+                            added_mass_coefficient=parachute.added_mass_coefficient: setattr(
+                                self,
+                                "parachute_added_mass_coefficient",
+                                added_mass_coefficient,
+                            ),
                         ]
                         self.flight_phases.add_phase(
                             node.t + parachute.lag,
@@ -1013,7 +1023,31 @@ class Flight:
                                             lambda self,
                                             parachute_cd_s=parachute.cd_s: setattr(
                                                 self, "parachute_cd_s", parachute_cd_s
-                                            )
+                                            ),
+                                            lambda self,
+                                            parachute_radius=parachute.radius: setattr(
+                                                self,
+                                                "parachute_radius",
+                                                parachute_radius,
+                                            ),
+                                            lambda self,
+                                            parachute_height=parachute.height: setattr(
+                                                self,
+                                                "parachute_height",
+                                                parachute_height,
+                                            ),
+                                            lambda self,
+                                            parachute_porosity=parachute.porosity: setattr(
+                                                self,
+                                                "parachute_porosity",
+                                                parachute_porosity,
+                                            ),
+                                            lambda self,
+                                            added_mass_coefficient=parachute.added_mass_coefficient: setattr(
+                                                self,
+                                                "parachute_added_mass_coefficient",
+                                                added_mass_coefficient,
+                                            ),
                                         ]
                                         self.flight_phases.add_phase(
                                             overshootable_node.t + parachute.lag,
@@ -1096,13 +1130,13 @@ class Flight:
         self.out_of_rail_time_index = 0
         self.out_of_rail_state = np.array([0])
         self.apogee_state = np.array([0])
+        self.apogee = 0
         self.apogee_time = 0
         self.x_impact = 0
         self.y_impact = 0
         self.impact_velocity = 0
         self.impact_state = np.array([0])
         self.parachute_events = []
-        self.post_processed = False
         self.__post_processed_variables = []
 
     def __init_flight_state(self):
@@ -1171,6 +1205,7 @@ class Flight:
             self.out_of_rail_state = self.initial_solution[1:]
             self.out_of_rail_time = self.initial_solution[0]
             self.out_of_rail_time_index = 0
+            self.t_initial = self.initial_solution[0]
             self.initial_derivative = self.u_dot_generalized
         if self._controllers or self.sensors:
             # Handle post process during simulation, get initial accel/forces
@@ -1679,6 +1714,12 @@ class Flight:
         ax, ay, az = K @ Vector(L)
         az -= self.env.gravity.get_value_opt(z)  # Include gravity
 
+        # Coriolis acceleration
+        _, w_earth_y, w_earth_z = self.env.earth_rotation_vector
+        ax -= 2 * (vz * w_earth_y - vy * w_earth_z)
+        ay -= 2 * (vx * w_earth_z)
+        az -= 2 * (-vx * w_earth_y)
+
         # Create u_dot
         u_dot = [
             vx,
@@ -1745,7 +1786,7 @@ class Flight:
         _, _, z, vx, vy, vz, e0, e1, e2, e3, omega1, omega2, omega3 = u
 
         # Create necessary vectors
-        # r = Vector([x, y, z])               # CDM position vector
+        # r = Vector([x, y, z])  # CDM position vector
         v = Vector([vx, vy, vz])  # CDM velocity vector
         e = [e0, e1, e2, e3]  # Euler parameters/quaternions
         w = Vector([omega1, omega2, omega3])  # Angular velocity vector
@@ -1904,9 +1945,6 @@ class Flight:
         # Angular velocity derivative
         w_dot = I_CM.inverse @ (T21 + (T20 ^ r_CM))
 
-        # Velocity vector derivative
-        v_dot = K @ (T20 / total_mass - (r_CM ^ w_dot))
-
         # Euler parameters derivative
         e_dot = [
             0.5 * (-omega1 * e1 - omega2 * e2 - omega3 * e3),
@@ -1914,6 +1952,10 @@ class Flight:
             0.5 * (omega2 * e0 - omega3 * e1 + omega1 * e3),
             0.5 * (omega3 * e0 + omega2 * e1 - omega1 * e2),
         ]
+
+        # Velocity vector derivative + Coriolis acceleration
+        w_earth = Vector(self.env.earth_rotation_vector)
+        v_dot = K @ (T20 / total_mass - (r_CM ^ w_dot)) - 2 * (w_earth ^ v)
 
         # Position vector derivative
         r_dot = [vx, vy, vz]
@@ -1959,15 +2001,9 @@ class Flight:
         wind_velocity_x = self.env.wind_velocity_x.get_value_opt(z)
         wind_velocity_y = self.env.wind_velocity_y.get_value_opt(z)
 
-        # Get Parachute data
-        cd_s = self.parachute_cd_s
-
         # Get the mass of the rocket
         mp = self.rocket.dry_mass
 
-        # Define constants
-        ka = 1  # Added mass coefficient (depends on parachute's porosity)
-        R = 1.5  # Parachute radius
         # to = 1.2
         # eta = 1
         # Rdot = (6 * R * (1 - eta) / (1.2**6)) * (
@@ -1975,8 +2011,17 @@ class Flight:
         # )
         # Rdot = 0
 
+        # tf = 8 * nominal diameter / velocity at line stretch
+
         # Calculate added mass
-        ma = ka * rho * (4 / 3) * np.pi * R**3
+        ma = (
+            self.parachute_added_mass_coefficient
+            * rho
+            * (2 / 3)
+            * np.pi
+            * self.parachute_radius**2
+            * self.parachute_height
+        )
 
         # Calculate freestream speed
         freestream_x = vx - wind_velocity_x
@@ -1985,14 +2030,20 @@ class Flight:
         free_stream_speed = (freestream_x**2 + freestream_y**2 + freestream_z**2) ** 0.5
 
         # Determine drag force
-        pseudo_drag = -0.5 * rho * cd_s * free_stream_speed
+        pseudo_drag = -0.5 * rho * self.parachute_cd_s * free_stream_speed
         # pseudo_drag = pseudo_drag - ka * rho * 4 * np.pi * (R**2) * Rdot
-        Dx = pseudo_drag * freestream_x
+        Dx = pseudo_drag * freestream_x  # add eta efficiency for wake
         Dy = pseudo_drag * freestream_y
         Dz = pseudo_drag * freestream_z
         ax = Dx / (mp + ma)
         ay = Dy / (mp + ma)
-        az = (Dz - 9.8 * mp) / (mp + ma)
+        az = (Dz - mp * self.env.gravity.get_value_opt(z)) / (mp + ma)
+
+        # Add coriolis acceleration
+        _, w_earth_y, w_earth_z = self.env.earth_rotation_vector
+        ax -= 2 * (vz * w_earth_y - vy * w_earth_z)
+        ay -= 2 * (vx * w_earth_z)
+        az -= 2 * (-vx * w_earth_y)
 
         if post_processing:
             self.__post_processed_variables.append(
@@ -2070,7 +2121,7 @@ class Flight:
 
     @funcify_method("Time (s)", "Y (m)", "spline", "constant")
     def y(self):
-        """Rocket y position relative to the lauch pad as a Function of
+        """Rocket y position relative to the launch pad as a Function of
         time."""
         return self.solution_array[:, [0, 2]]
 
@@ -2966,7 +3017,7 @@ class Flight:
         "Time (s)", "Horizontal Distance to Launch Point (m)", "spline", "constant"
     )
     def drift(self):
-        """Rocket horizontal distance to tha launch point, in meters, as a
+        """Rocket horizontal distance to the launch point, in meters, as a
         Function of time."""
         return np.column_stack(
             (self.time, (self.x[:, 1] ** 2 + self.y[:, 1] ** 2) ** 0.5)
@@ -3055,14 +3106,16 @@ class Flight:
         null_force = Function(0)
         if self.out_of_rail_time_index == 0:  # No rail phase, no rail button forces
             warnings.warn(
-                "Trying to calculate rail button forces without a rail phase defined."
-                + "The rail button forces will be set to zero."
+                "Trying to calculate rail button forces without a rail phase defined. "
+                + "The rail button forces will be set to zero.",
+                UserWarning,
             )
             return null_force, null_force, null_force, null_force
         if len(self.rocket.rail_buttons) == 0:
             warnings.warn(
-                "Trying to calculate rail button forces without rail buttons defined."
-                + "The rail button forces will be set to zero."
+                "Trying to calculate rail button forces without rail buttons defined. "
+                + "The rail button forces will be set to zero.",
+                UserWarning,
             )
             return null_force, null_force, null_force, null_force
 
@@ -3174,28 +3227,6 @@ class Flight:
 
         return np.array(self.__post_processed_variables)
 
-    def post_process(self, interpolation="spline", extrapolation="natural"):
-        """This method is **deprecated** and is only kept here for backwards
-        compatibility. All attributes that need to be post processed are
-        computed just in time.
-
-        Post-process all Flight information produced during
-        simulation. Includes the calculation of maximum values,
-        calculation of secondary values such as energy and conversion
-        of lists to Function objects to facilitate plotting.
-
-        Returns
-        -------
-        None
-        """
-        # pylint: disable=unused-argument
-        warnings.warn(
-            "The method post_process is deprecated and will be removed in v1.10. "
-            "All attributes that need to be post processed are computed just in time.",
-            DeprecationWarning,
-        )
-        self.post_processed = True
-
     def calculate_stall_wind_velocity(self, stall_angle):  # TODO: move to utilities
         """Function to calculate the maximum wind velocity before the angle of
         attack exceeds a desired angle, at the instant of departing rail launch.
@@ -3235,191 +3266,53 @@ class Flight:
             + f" of attack exceeds {stall_angle:.3f}°: {w_v:.3f} m/s"
         )
 
-    def export_pressures(self, file_name, time_step):  # TODO: move out
-        """Exports the pressure experienced by the rocket during the flight to
-        an external file, the '.csv' format is recommended, as the columns will
-        be separated by commas. It can handle flights with or without
-        parachutes, although it is not possible to get a noisy pressure signal
-        if no parachute is added.
-
-        If a parachute is added, the file will contain 3 columns: time in
-        seconds, clean pressure in Pascals and noisy pressure in Pascals.
-        For flights without parachutes, the third column will be discarded
-
-        This function was created especially for the 'Projeto Jupiter'
-        Electronics Subsystems team and aims to help in configuring
-        micro-controllers.
-
-        Parameters
-        ----------
-        file_name : string
-            The final file name,
-        time_step : float
-            Time step desired for the final file
-
-        Return
-        ------
-        None
+    @deprecated(
+        reason="Moved to FlightDataExporter.export_pressures()",
+        version="v1.12.0",
+        alternative="rocketpy.simulation.flight_data_exporter.FlightDataExporter.export_pressures",
+    )
+    def export_pressures(self, file_name, time_step):
         """
-        time_points = np.arange(0, self.t_final, time_step)
-        # pylint: disable=W1514, E1121
-        with open(file_name, "w") as file:
-            if len(self.rocket.parachutes) == 0:
-                print("No parachutes in the rocket, saving static pressure.")
-                for t in time_points:
-                    file.write(f"{t:f}, {self.pressure.get_value_opt(t):.5f}\n")
-            else:
-                for parachute in self.rocket.parachutes:
-                    for t in time_points:
-                        p_cl = parachute.clean_pressure_signal_function.get_value_opt(t)
-                        p_ns = parachute.noisy_pressure_signal_function.get_value_opt(t)
-                        file.write(f"{t:f}, {p_cl:.5f}, {p_ns:.5f}\n")
-                    # We need to save only 1 parachute data
-                    break
+        .. deprecated:: 1.11
+           Use :class:`rocketpy.simulation.flight_data_exporter.FlightDataExporter`
+           and call ``.export_pressures(...)``.
+        """
+        return FlightDataExporter(self).export_pressures(file_name, time_step)
 
+    @deprecated(
+        reason="Moved to FlightDataExporter.export_data()",
+        version="v1.12.0",
+        alternative="rocketpy.simulation.flight_data_exporter.FlightDataExporter.export_data",
+    )
     def export_data(self, file_name, *variables, time_step=None):
-        """Exports flight data to a comma separated value file (.csv).
-
-        Data is exported in columns, with the first column representing time
-        steps. The first line of the file is a header line, specifying the
-        meaning of each column and its units.
-
-        Parameters
-        ----------
-        file_name : string
-            The file name or path of the exported file. Example: flight_data.csv
-            Do not use forbidden characters, such as / in Linux/Unix and
-            `<, >, :, ", /, \\, | ?, *` in Windows.
-        variables : strings, optional
-            Names of the data variables which shall be exported. Must be Flight
-            class attributes which are instances of the Function class. Usage
-            example: test_flight.export_data('test.csv', 'z', 'angle_of_attack',
-            'mach_number').
-        time_step : float, optional
-            Time step desired for the data. If None, all integration time steps
-            will be exported. Otherwise, linear interpolation is carried out to
-            calculate values at the desired time steps. Example: 0.001.
         """
-        # TODO: we should move this method to outside of class.
-
-        # Fast evaluation for the most basic scenario
-        if time_step is None and len(variables) == 0:
-            np.savetxt(
-                file_name,
-                self.solution,
-                fmt="%.6f",
-                delimiter=",",
-                header=""
-                "Time (s),"
-                "X (m),"
-                "Y (m),"
-                "Z (m),"
-                "E0,"
-                "E1,"
-                "E2,"
-                "E3,"
-                "W1 (rad/s),"
-                "W2 (rad/s),"
-                "W3 (rad/s)",
-            )
-            return
-
-        # Not so fast evaluation for general case
-        if variables is None:
-            variables = [
-                "x",
-                "y",
-                "z",
-                "vx",
-                "vy",
-                "vz",
-                "e0",
-                "e1",
-                "e2",
-                "e3",
-                "w1",
-                "w2",
-                "w3",
-            ]
-
-        if time_step is None:
-            time_points = self.time
-        else:
-            time_points = np.arange(self.t_initial, self.t_final, time_step)
-
-        exported_matrix = [time_points]
-        exported_header = "Time (s)"
-
-        # Loop through variables, get points and names (for the header)
-        for variable in variables:
-            if variable in self.__dict__:
-                variable_function = self.__dict__[variable]
-            # Deal with decorated Flight methods
-            else:
-                try:
-                    obj = getattr(self.__class__, variable)
-                    variable_function = obj.__get__(self, self.__class__)
-                except AttributeError as exc:
-                    raise AttributeError(
-                        f"Variable '{variable}' not found in Flight class"
-                    ) from exc
-            variable_points = variable_function(time_points)
-            exported_matrix += [variable_points]
-            exported_header += f", {variable_function.__outputs__[0]}"
-
-        exported_matrix = np.array(exported_matrix).T  # Fix matrix orientation
-
-        np.savetxt(
-            file_name,
-            exported_matrix,
-            fmt="%.6f",
-            delimiter=",",
-            header=exported_header,
-            encoding="utf-8",
+        .. deprecated:: 1.11
+           Use :class:`rocketpy.simulation.flight_data_exporter.FlightDataExporter`
+           and call ``.export_data(...)``.
+        """
+        return FlightDataExporter(self).export_data(
+            file_name, *variables, time_step=time_step
         )
 
+    @deprecated(
+        reason="Moved to FlightDataExporter.export_sensor_data()",
+        version="v1.12.0",
+        alternative="rocketpy.simulation.flight_data_exporter.FlightDataExporter.export_sensor_data",
+    )
     def export_sensor_data(self, file_name, sensor=None):
-        """Exports sensors data to a file. The file format can be either .csv or
-        .json.
-
-        Parameters
-        ----------
-        file_name : str
-            The file name or path of the exported file. Example: flight_data.csv
-            Do not use forbidden characters, such as / in Linux/Unix and
-            `<, >, :, ", /, \\, | ?, *` in Windows.
-        sensor : Sensor, string, optional
-            The sensor to export data from. Can be given as a Sensor object or
-            as a string with the sensor name. If None, all sensors data will be
-            exported. Default is None.
         """
-        if sensor is None:
-            data_dict = {}
-            for used_sensor, measured_data in self.sensor_data.items():
-                data_dict[used_sensor.name] = measured_data
-        else:
-            # export data of only that sensor
-            data_dict = {}
+        .. deprecated:: 1.11
+           Use :class:`rocketpy.simulation.flight_data_exporter.FlightDataExporter`
+           and call ``.export_sensor_data(...)``.
+        """
+        return FlightDataExporter(self).export_sensor_data(file_name, sensor=sensor)
 
-            if not isinstance(sensor, str):
-                data_dict[sensor.name] = self.sensor_data[sensor]
-            else:  # sensor is a string
-                matching_sensors = [s for s in self.sensor_data if s.name == sensor]
-
-                if len(matching_sensors) > 1:
-                    data_dict[sensor] = []
-                    for s in matching_sensors:
-                        data_dict[s.name].append(self.sensor_data[s])
-                elif len(matching_sensors) == 1:
-                    data_dict[sensor] = self.sensor_data[matching_sensors[0]]
-                else:
-                    raise ValueError("Sensor not found in the Flight.sensor_data.")
-
-        with open(file_name, "w") as file:
-            json.dump(data_dict, file)
-        print("Sensor data exported to: ", file_name)
-
-    def export_kml(  # TODO: should be moved out of this class.
+    @deprecated(
+        reason="Moved to FlightDataExporter.export_kml()",
+        version="v1.12.0",
+        alternative="rocketpy.simulation.flight_data_exporter.FlightDataExporter.export_kml",
+    )
+    def export_kml(
         self,
         file_name="trajectory.kml",
         time_step=None,
@@ -3427,78 +3320,18 @@ class Flight:
         color="641400F0",
         altitude_mode="absolute",
     ):
-        """Exports flight data to a .kml file, which can be opened with Google
-        Earth to display the rocket's trajectory.
-
-        Parameters
-        ----------
-        file_name : string
-            The file name or path of the exported file. Example: flight_data.csv
-        time_step : float, optional
-            Time step desired for the data. If None, all integration time steps
-            will be exported. Otherwise, linear interpolation is carried out to
-            calculate values at the desired time steps. Example: 0.001.
-        extrude: bool, optional
-            To be used if you want to project the path over ground by using an
-            extruded polygon. In case False only the linestring containing the
-            flight path will be created. Default is True.
-        color : str, optional
-            Color of your trajectory path, need to be used in specific kml
-            format. Refer to http://www.zonums.com/gmaps/kml_color/ for more
-            info.
-        altitude_mode: str
-            Select elevation values format to be used on the kml file. Use
-            'relativetoground' if you want use Above Ground Level elevation, or
-            'absolute' if you want to parse elevation using Above Sea Level.
-            Default is 'relativetoground'. Only works properly if the ground
-            level is flat. Change to 'absolute' if the terrain is to irregular
-            or contains mountains.
         """
-        # Define time points vector
-        if time_step is None:
-            time_points = self.time
-        else:
-            time_points = np.arange(self.t_initial, self.t_final + time_step, time_step)
-        # Open kml file with simplekml library
-        kml = simplekml.Kml(open=1)
-        trajectory = kml.newlinestring(name="Rocket Trajectory - Powered by RocketPy")
-
-        if altitude_mode == "relativetoground":
-            # In this mode the elevation data will be the Above Ground Level
-            # elevation. Only works properly if the ground level is similar to
-            # a plane, i.e. it might not work well if the terrain has mountains
-            coords = [
-                (
-                    self.longitude.get_value_opt(t),
-                    self.latitude.get_value_opt(t),
-                    self.altitude.get_value_opt(t),
-                )
-                for t in time_points
-            ]
-            trajectory.coords = coords
-            trajectory.altitudemode = simplekml.AltitudeMode.relativetoground
-        else:  # altitude_mode == 'absolute'
-            # In this case the elevation data will be the Above Sea Level elevation
-            # Ensure you use the correct value on self.env.elevation, otherwise
-            # the trajectory path can be offset from ground
-            coords = [
-                (
-                    self.longitude.get_value_opt(t),
-                    self.latitude.get_value_opt(t),
-                    self.z.get_value_opt(t),
-                )
-                for t in time_points
-            ]
-            trajectory.coords = coords
-            trajectory.altitudemode = simplekml.AltitudeMode.absolute
-        # Modify style of trajectory linestring
-        trajectory.style.linestyle.color = color
-        trajectory.style.polystyle.color = color
-        if extrude:
-            trajectory.extrude = 1
-        # Save the KML
-        kml.save(file_name)
-        print("File ", file_name, " saved with success!")
+        .. deprecated:: 1.11
+           Use :class:`rocketpy.simulation.flight_data_exporter.FlightDataExporter`
+           and call ``.export_kml(...)``.
+        """
+        return FlightDataExporter(self).export_kml(
+            file_name=file_name,
+            time_step=time_step,
+            extrude=extrude,
+            color=color,
+            altitude_mode=altitude_mode,
+        )
 
     def info(self):
         """Prints out a summary of the data available about the Flight."""
@@ -3515,7 +3348,7 @@ class Flight:
             yield i, node_list[i]
             i += 1
 
-    def to_dict(self, include_outputs=False):
+    def to_dict(self, **kwargs):
         data = {
             "rocket": self.rocket,
             "env": self.env,
@@ -3544,7 +3377,6 @@ class Flight:
             "x_impact": self.x_impact,
             "y_impact": self.y_impact,
             "t_final": self.t_final,
-            "flight_phases": self.flight_phases,
             "function_evaluations": self.function_evaluations,
             "ax": self.ax,
             "ay": self.ay,
@@ -3558,9 +3390,10 @@ class Flight:
             "M1": self.M1,
             "M2": self.M2,
             "M3": self.M3,
+            "net_thrust": self.net_thrust,
         }
 
-        if include_outputs:
+        if kwargs.get("include_outputs", False):
             data.update(
                 {
                     "time": self.time,